feat(web): extract quality dashboard helpers into quality-dashboard-lib
